demo: Fix color scaling behavior for simplification #nojira
When simplifying meshes, we no longer need to divide by 255 because
BufferAttribute.getX will do this for us as of three.js 144. This was
missed during a three.js version upgrade which broke attribute awareness
in the demo by using very small color values that ended up not affecting
the simplification result.

When simplifying points, we are converting typed arrays without any
normalization; as such, we need to normalize these by applying the
scaling factor to color weight.