gltfpack: Add fuzzing dictionary and seed GLB file
Note that the dictionary doesn't contain a lot of glTF extensions
but it's not clear if we need this, because LLVM libFuzzer uses a
mutator that uses strings from the binary in addition to supplied
dictionary data.

The GLB seed is taken from glTF-Sample-Assets BoxTextured, and is
simultaneously reasonably small and covers all the basic features so
it's a good candidate to use for a seed.