gltfpack: Focus fuzzing coverage on valid GLB files
To make sure that we don't spend too much time fuzzing the cgltf parser,
we only add the files to the corpus if the parsing succeeded.

As the comment explains, this is a tradeoff and it's not fully clear
what is optimal here, but for now this seems to work okay.