clusterizer: Switch to a small cache for optimizeMeshlet
Instead of only matching against the last triangle, match all previous
triangles by tracking the delta between a given cache position; the
delta is computed using 8-bit unsigned math so that overflows don't
matter as much (there's still a collision between triangles 0 and 256 in
the cache but that shouldn't affect the quality noticeably).

Experimentally, a cutoff of 3 or 4 produce the same results on NVidia so
use 3 for more conservative matching.