gltfpack: Preserve image names in the output
Image names can be useful for identifying images in the output file when
embedded images are used, or for custom application processing e.g. for
attaching extra images to materials via extra properties.

Note that there's redundancy between images and textures; for now we choose
to preserve image names and not texture names, as this seems to match
conventions used by Blender glTF export.