Create USERS.md (#687)
This is an initial version of a list of known projects that are using meshoptimizer. It's missing a lot of projects I've been told about and don't remember OTOH, and it's likely missing a *lot* of games that use it that I don't know about.

The intent here is to slowly expand this list over time with prominent projects (commercial software or open source software).