gltfpack: Use vertex color information for simplification
When -sv command line argument is specified, we will use vertex colors
(if present) as part of attribute aware simplification to produce better
results. For now the intention is to use a single argument to
incorporate multiple different types of attributes with fixed weights,
in the future we might expose more detailed control via '-sv?' prefix.