gltfpack: Allow simplification with target vertex/index count 0
The existing check results in an odd discontinuity: eg when simplifying
with ratio 0.1, a 10-triangle mesh will be simplified to a target of 1,
but a 9-triangle mesh will be left alone. This becomes more and more
problematic for higher simplification ratios, where 0.01 ratio will
leave small meshes completely unsimplified.

In the future we probably need a more careful simplification schedule;
however, this logic is currently extraneous as we will remove empty
meshes after processing so we shouldn't need any early outs here, and
for triangle simplification the error bounds should preserve detail.