gltfpack: Adjust handling of textures with redundant images
When a texture has both source and encoded image, we prefer to extract
various metadata from the source image, but preserve both when writing
the texture object.

When -tc is used, we encode the source image and leave the
pre-compressed image unused (this is a little redundant but requires
unused image filtering to fix).