gltfpack: Preserve mesh attributes if the texture exists with no image
There are some cases when the texture refers to a NULL image, eg when an
unknown image extension is used such as EXT_texture_webp. In this case
we used to remove UV coordinates which triggered excessive validation
errors; we now preserve UV coordinates even though the texture object
is not fully serialized/processed.