gltfpack: Remove generation of meshlet bounds meshes
The bounding spheres here hurt more than they help: due to the lack of a
material by default they obscure the actual meshlets, and if the
importer merges the meshes on import then these become hard to separate.
Additionally, the spheres mostly don't help with debugging meshlet
clustering.

It could make sense to switch to a mesh per meshet for this
visualization in the future if we decide having a sense of meshlet size
is important, but for now the pure visualization is more helpful.