simplify: Implement initial support for sparse simplification
When simplifying a small subset of the larger mesh, all computations
that go over the entire vertex buffer become expensive; this adds up
even when done once, and especially when done for every pass.

This change introduces a sparse simplification mode that instructs the
simplifier to optimize based on the assumption that the subset of the
mesh that is being simplified is small. In that case it's worth spending
extra time to convert indices into a small 0..U subrange, do all
internal processing assuming we are working with a small vertex/index
buffer, and remap the indices at the end. While this processing could be
done externally, that is less efficient as it requires constant copying
of position/attribute data; in constrast, we can do it fairly cheaply.

We need to take sparse_remap into account for any code that indexes
input position/attribute data; buildPositionRemap needs this as well as
it currently works off of original (unscaled) data.

buildSparseRemap does need to perform O(dense) work if we want to avoid
using hash maps for deduplication/filtering; this change uses a small
zeroed array and a large sparsely-initialized array to reduce the cost.
This is subject to future improvements (although is fairly performant as
it is when simplifying 500-triangle subsets of a 800K triangle mesh).