simplify: Use a bit set for sparse remap computation
We have two parts of buildSparseRemap that are still dependent on the
number of vertices wrt complexity, filter clearing and revremap
allocation. We could replace revremap with a hash map if a large
allocation proves to be problematic, but it also might work fine in
practice - whereas filter[] clearing is a cost we must pay and for small
subsets of large meshes this can be 20% of the entire simplification.

To fix this we now use a bit set which is 8x cheaper to clear (the
actual addressing gets more expensive but it's a fraction of the cost
due to sparsity assumption).