demo: Add a test for sparse simplification
The test needs to check that positions, attributes and lock flags are
all addressed using proper indexing. We do assume that collapses are
going in a specific direction (the input data technically permits two
collapse directions for each test), if this becomes a problem due to fp
instabilities we can tweak the input data then.