simplify: Use a hash map to generate sparse remap
Instead of using a large uninitialized allocation, we now use a hash
table. Under the assumption that a mesh subset is much smaller, we were
relying on the efficiency of repeatedly reallocating a large
uninitialized segment of memory, which worked well on some systems
(Linux) but not as much on others (Windows). Instead, we now minimize
the allocation size.

This ends up a little slower when the sparsity assumption does not hold,
as we no longer use direct indexing. To minimize the impact, we use a
hash function that reduces the amount of avalanche so that sequential
indices have fewer hash conflicts; it might also make sense to use an
identity function here, or a small multiplicative constant.

To minimize the impact further, we don't store pairs in the map, and
just store the new index; when doing a lookup, due to the unique
construction of the hasher we can lookup the dense index despite the
fact that the map only stores sparse indices.