gltfpack: Rebalance texture quality table
The main goal of this change is to tweak the UASTC RDO curve. The
initial values here have been picked for early UASTC versions that used
a different RDO algorithm, and haven't been revisited since.

The new table makes it so that at tq 5, we use the same defaults as
basisu (q=128 for ETC1 and l=1 for UASTC RDO), and tweak the table to be
in general a little more aggressive for UASTC. We use different curves
before l=1 and after as the expectation is that for earlier quality
levels, the size is much more important than quality.

The default lambda for tq=8 is a little larger now (0.4 vs 0.3) which
will result in a small size improvement and still good quality by
default. There's maybe a little more room for default lambda to be
higher, but eg lambda=1 does not seem practical for normal maps on a
range of production assets, so we keep the curve for tq=6..10 to be more
quality focused.

The new table is not based on a rigorous fitting of the actual
quality/size curves, but the new tool gltfbasis.py can help find a more
optimal table in the future.