simplify: Make hasTriangleFlip predicate use an angle cutoff
Instead of checking if the dot product is 0 or negative (90 degree limit
on triangle rotations), use a cutoff of ~75 degrees. The motivation here
is that meshes with complex curvature can end up flipping triangles
through a series of collapses, and since we always evaluate flips
compared to the current mesh we miss that. Technically this can happen
even with a 75 degree limit, but in practice this solves cases of flips
that have been reported before, and doesn't noticeably constrain
simplification or slow it down.

For numerical stability, we avoid dividing by the triangle areas (as
they can be close to zero) and instead use the areas to compute a
bound for dot product. Even if the area product underflows, we just get
the same check that we used to have -- areas should be in [0..1] so
there should be no risk of overflow.