js: Adjust simplify/simplifyWithAttributes assertions
To make ErrorAbsolute mode useful in JS, it needs to allow values
outside of 0..1 range; for simplicity we just remove the upper bound for
now.

Also, previously flags could contain values that are not recognized as
valid options and this would silently work; these are now checked with
assertions to prevent mistakes.