Add make symbols to show the shared object export table
This is convenient to make sure symbols aren't acciendtally exposed
unnecessarily. Note that for now this builds the code with default
visibility settings, instead of using special shared object setup we
have in CMake, so in debug mode there will be extra inline symbols.

Also this needs to build a .so with special flags so we do this via an
amalgamated source file which helpfully tests that amalgamation is
actually supported.