gltfpack: Switch to SimplifyInternalDebug for debugSimplify
This code reworks the debug visualization for simplification to use new
internal feature that communicates metadata via index data.

For now we draw loops even for locked vertices and don't make an attempt
to deduplicate points or lines, as visualization performance is not
critical, and locked loops actually make it easier to debug in some
cases.